requirements.txt
fastapi
websockets
uvicorn
python-dotenv